Fix the card counter Postgres refused to plan
One placeholder fed both card_points (bigint) and best_drop_points (integer),
and Postgres would not deduce a single type for it: "inconsistent types deduced
for parameter $3". Every card a player pulled was dropped.

The design held — the error went to the log and the roll itself was unaffected —
which is also why nothing looked broken until the counters were read. Found by
running it against a real database; casting both uses fixes it.

+++ internal/repository/seasons.go
@@ -125,7 +125,7 @@ func (r *PostgresRepo) AddSeasonCard(seasonID int, userID int64, cardID, points 
 	_, err := r.db.Exec(`
 		INSERT INTO season_stats (season_id, user_id, card_points, cards_count,
 		                          best_drop_card_id, best_drop_points, best_drop_at)
-		VALUES ($1, $2, $3, 1, $4, $3, NOW())
+		VALUES ($1, $2, $3::bigint, 1, $4, $3::integer, NOW())
 		ON CONFLICT (season_id, user_id) DO UPDATE
 		SET card_points = season_stats.card_points + EXCLUDED.card_points,
 		    cards_count = season_stats.cards_count + 1,
